refactor: TableGroup 생성 검증 기능 이동
