+++ src/main/java/kitchenpos/ordertable/domain/TableGroup.java
@@ -1,21 +1,14 @@
 package kitchenpos.ordertable.domain;
 
-import static java.util.stream.Collectors.toSet;
 import static javax.persistence.GenerationType.IDENTITY;
-import static kitchenpos.exception.ExceptionType.ALREADY_ASSIGNED_TABLE_GROUP;
-import static kitchenpos.exception.ExceptionType.DUPLICATED_TABLES_OF_TABLE_GROUP;
-import static kitchenpos.exception.ExceptionType.INVALID_TABLES_COUNT_OF_TABLE_GROUP;
-import static kitchenpos.exception.ExceptionType.NOT_EMPTY_ORDER_TABLE_IN_TABLE_GROUP;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Set;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.OneToMany;
-import kitchenpos.exception.CustomException;
 
 @Entity
 public class TableGroup {
@@ -32,47 +25,14 @@ public class TableGroup {
     }
 
     public TableGroup(Long id, LocalDateTime createdDate, List<OrderTable> orderTables) {
-        validateOrderTables(orderTables);
         this.id = id;
         this.createdDate = createdDate;
         this.orderTables = orderTables;
+    }
+
+    public void group() {
         for (OrderTable orderTable : orderTables) {
             orderTable.group(id);
-        }
-    }
-
-    private void validateOrderTables(List<OrderTable> orderTables) {
-        validateOrderTablesSize(orderTables);
-        validateOrderTablesDistinct(orderTables);
-        validateOrderTablesHaveNoGroup(orderTables);
-        validateOrderTablesEmpty(orderTables);
-    }
-
-    private void validateOrderTablesSize(List<OrderTable> orderTables) {
-        if (orderTables == null || orderTables.size() < 2) {
-            throw new CustomException(INVALID_TABLES_COUNT_OF_TABLE_GROUP);
-        }
-    }
-
-    private void validateOrderTablesDistinct(List<OrderTable> orderTables) {
-        Set<Long> uniqueOrderTableIds = orderTables.stream()
-                                                   .map(OrderTable::getId)
-                                                   .collect(toSet());
-
-        if (orderTables.size() != uniqueOrderTableIds.size()) {
-            throw new CustomException(DUPLICATED_TABLES_OF_TABLE_GROUP);
-        }
-    }
-
-    private void validateOrderTablesHaveNoGroup(List<OrderTable> orderTables) {
-        if (orderTables.stream().anyMatch(orderTable -> orderTable.getTableGroupId() != null)) {
-            throw new CustomException(ALREADY_ASSIGNED_TABLE_GROUP);
-        }
-    }
-
-    private void validateOrderTablesEmpty(List<OrderTable> orderTables) {
-        if (orderTables.stream().anyMatch(orderTable -> !orderTable.isEmpty())) {
-            throw new CustomException(NOT_EMPTY_ORDER_TABLE_IN_TABLE_GROUP);
         }
     }
