+++ src/main/java/kitchenpos/ordertable/service/TableGroupMapper.java
@@ -28,6 +28,5 @@ public class TableGroupMapper {
             .createdDate(LocalDateTime.now())
             .orderTables(orderTables)
             .build();
-
     }
 }
